Adding django-sphinxdoc support.
For guide on how to generate the documentation see doc/readme.txt.
Currently doc generation is not hooked up in fabric and will not be
done automatically. Moreover apidoc if now empty, but will be fixed
soon together with a correct project layout.

In addition to the new features, admin paneli was fixe and should be
now correctly accessible under 'admin/'.

Django-sphinxdoc requires haystack of version <2.0 to be included in the
project. Currently, the search is using 'dummy' backend, but should be
configured correctly in the near future.