New project layout.
All files directly related to the webapp were moved up a folder to
app/, so did manage.py. Fabfile was changed to reflect this and the
default target_def requires 2 more keys configured.

Static and template folders are now in root of the repository along
with doc, bin and deployment.