MTRACKER-12: Autodoc integration, deployment.
Autodoc documentation should now build and be correctly included in
the admin. Two modules which had code executing on import were
altered to allow doc generating.

Docs generating with autodoc + django-sphinxdoc is now included in the
deployment process. The documentation is copied to a separate folder in
the project directory and built there.